update tableView Scroll mechanism
